chips: rp2xxx: share the GPIO pad control enums
chips/rp2040/src/gpio.rs and chips/rp2350/src/gpio.rs each need a
SlewRate and a DriveStrength to describe GPIO_PAD, and the two
declarations are identical character for character. Copying the second
one into place is what @ppannuto asked about on #5126: they should be in
the shared crate and re-exported instead.

They can be, and unlike the rest of GPIO they are not a judgement call.
The two chips' GPIO blocks diverge by hundreds of lines and stay where
they are, but GPIO_PAD does not diverge at all: DRIVE sits at bits 5:4
on both, SCHMITT at bit 1, SLEWFAST at bit 0, and drive strength encodes
2, 4, 8 and 12 mA as 0 to 3 on both. The enums describe those fields and
nothing else, so sharing them shares a fact about the silicon rather
than a guess that two things will stay alike.

So chips/rp2xxx gains src/pads.rs holding both, and rp2040 re-exports
them alongside its other imports. rp2350 follows in the next commit,
which is what needs them.

The shared copy is the RP2040's, with one change: SlewRate declared Fast
before Slow, which reads backwards for a two variant enum whose
discriminants are 1 and 0. The variants are ordered by discriminant here.
Both carry explicit values and are written with `as u32`, so ordering
them cannot change what any caller does.

DriveStrength keeps this file's inconsistent spelling: Drive2mA with a
capital A, then Drive4ma, Drive8ma and Drive12ma without. Correcting it
would touch chips/rp2040/src/pio.rs, the only caller of either enum, and
that is the file #5126 is currently moving code out of. Two open PRs
should not both be editing it. Worth doing once #5126 lands.

All five RP2040 and RP2350 boards are unchanged, and checked per section
rather than by the summary column, which pads to a 4K boundary and can
hide a change smaller than a page:

    board                  text   data     bss
    raspberry_pi_pico     98348      0   16528
    pico_explorer_base   102444      0   72668
    nano_rp2040_connect   94252      0   14368
    raspberry_pi_pico_w  352300      0   20724
    raspberry_pi_pico_2   69164      0   19100

On raspberry_pi_pico_w, the only board that calls either enum, .text is
349532 and .storage 2452 both before and after.
